move sparse utilities to lenskit.math.sparse
