+++ lenskit/lenskit/algorithms/knn/item.py
@@ -23,9 +23,9 @@ from typing_extensions import Callable, Literal, Optional, TypeAlias, override
 from lenskit import util
 from lenskit.data import FeedbackType
 from lenskit.data.dataset import Dataset
-from lenskit.data.matrix import normalize_sparse_rows, safe_spmv
 from lenskit.data.vocab import Vocabulary
 from lenskit.diagnostics import ConfigWarning, DataWarning
+from lenskit.math.sparse import normalize_sparse_rows, safe_spmv
 from lenskit.parallel import ensure_parallel_init
 from lenskit.util.logging import pbh_update, progress_handle
