+++ lenskit/lenskit/data/matrix.py
@@ -12,13 +12,12 @@ Data manipulation routines.
 from __future__ import annotations
 
 import logging
-import platform
 
 import numpy as np
 import scipy.sparse as sps
 import torch
 from numpy.typing import ArrayLike
-from typing_extensions import Literal, NamedTuple, Optional, TypeVar, overload
+from typing_extensions import NamedTuple, Optional, TypeVar
 
 _log = logging.getLogger(__name__)
 
@@ -111,120 +110,3 @@ class InteractionMatrix:
         The shape of the interaction matrix (rows x columns).
         """
         return (self.n_users, self.n_items)
-
-
-@overload
-def normalize_sparse_rows(
-    matrix: t.Tensor, method: Literal["center"], inplace: bool = False
-) -> tuple[t.Tensor, t.Tensor]: ...
-@overload
-def normalize_sparse_rows(
-    matrix: t.Tensor, method: Literal["unit"], inplace: bool = False
-) -> tuple[t.Tensor, t.Tensor]: ...
-def normalize_sparse_rows(
-    matrix: t.Tensor, method: str, inplace: bool = False
-) -> tuple[t.Tensor, t.Tensor]:
-    """
-    Normalize the rows of a sparse matrix.
-    """
-    match method:
-        case "unit":
-            return _nsr_unit(matrix)
-        case "center":
-            return _nsr_mean_center(matrix)
-        case _:
-            raise ValueError(f"unsupported normalization method {method}")
-
-
-def _nsr_mean_center(matrix: t.Tensor) -> tuple[t.Tensor, t.Tensor]:
-    nr, _nc = matrix.shape
-    sums = matrix.sum(dim=1, keepdim=True).to_dense().reshape(nr)
-    counts = torch.diff(matrix.crow_indices())
-    assert sums.shape == counts.shape
-    means = torch.nan_to_num(sums / counts, 0)
-    return t.sparse_csr_tensor(
-        crow_indices=matrix.crow_indices(),
-        col_indices=matrix.col_indices(),
-        values=matrix.values() - t.repeat_interleave(means, counts),
-        size=matrix.shape,
-    ), means
-
-
-def _nsr_unit(matrix: t.Tensor) -> tuple[t.Tensor, t.Tensor]:
-    sqmat = t.sparse_csr_tensor(
-        crow_indices=matrix.crow_indices(),
-        col_indices=matrix.col_indices(),
-        values=matrix.values().square(),
-    )
-    norms = sqmat.sum(dim=1, keepdim=True).to_dense().reshape(matrix.shape[0])
-    norms.sqrt_()
-    recip_norms = t.where(norms > 0, t.reciprocal(norms), 0.0)
-    return t.sparse_csr_tensor(
-        crow_indices=matrix.crow_indices(),
-        col_indices=matrix.col_indices(),
-        values=matrix.values() * t.repeat_interleave(recip_norms, matrix.crow_indices().diff()),
-        size=matrix.shape,
-    ), norms
-
-
-def torch_sparse_from_scipy(
-    M: sps.coo_array, layout: Literal["csr", "coo", "csc"] = "coo"
-) -> t.Tensor:
-    """
-    Convert a SciPy :class:`sps.coo_array` into a torch sparse tensor.
-    """
-    ris = t.from_numpy(M.row)
-    cis = t.from_numpy(M.col)
-    vs = t.from_numpy(M.data)
-    indices = t.stack([ris, cis])
-    assert indices.shape == (2, M.nnz)
-    T = t.sparse_coo_tensor(indices, vs, size=M.shape)
-    assert T.shape == M.shape
-
-    match layout:
-        case "csr":
-            return T.to_sparse_csr()
-        case "csc":
-            return T.to_sparse_csc()
-        case "coo":
-            return T.coalesce()
-        case _:
-            raise ValueError(f"invalid layout {layout}")
-
-
-if platform.machine() == "arm64":
-
-    @torch.jit.ignore  # type: ignore
-    def safe_spmv(matrix, vector):  # type: ignore
-        """
-        Sparse matrix-vector multiplication working around PyTorch bugs.
-
-        This is equivalent to :func:`torch.mv` for sparse CSR matrix
-        and dense vector, but it works around PyTorch bug 127491_ by
-        falling back to SciPy on ARM.
-
-        .. _127491: https://github.com/pytorch/pytorch/issues/127491
-        """
-        assert matrix.is_sparse_csr
-        nr, nc = matrix.shape
-        M = sps.csr_array(
-            (matrix.values().numpy(), matrix.col_indices().numpy(), matrix.crow_indices().numpy()),
-            (nr, nc),
-        )
-        v = vector.numpy()
-        return torch.from_numpy(M @ v)
-
-else:
-
-    def safe_spmv(matrix: torch.Tensor, vector: torch.Tensor) -> torch.Tensor:
-        """
-        Sparse matrix-vector multiplication working around PyTorch bugs.
-
-        This is equivalent to :func:`torch.mv` for sparse CSR matrix
-        and dense vector, but it works around PyTorch bug 127491_ by
-        falling back to SciPy on ARM.
-
-        .. _127491: https://github.com/pytorch/pytorch/issues/127491
-        """
-        assert matrix.is_sparse_csr
-        return torch.mv(matrix, vector)
